# backend/api/openapi/swagger.yaml
definitions:
...
  sico-backend_internal_transport_http_dto_llmhubs.AuthConfig:
    properties:
      apiKeyValue:
        type: string
      authType:
        type: integer
      headerName:
        type: string
      token:
        type: string
    type: object
  sico-backend_internal_transport_http_dto_llmhubs.CreateModelRegistryData:
    properties:
      model:
        $ref: '#/definitions/sico-backend_internal_transport_http_dto_llmhubs.ModelRegistryEntry'
    type: object
  sico-backend_internal_transport_http_dto_llmhubs.CreateModelRegistryRequest:
    properties:
      agentId:
        type: string
      auth:
        $ref: '#/definitions/sico-backend_internal_transport_http_dto_llmhubs.AuthConfig'
      config:
        $ref: '#/definitions/structpb.Struct'
      description:
        type: string
      displayName:
        maxLength: 255
        type: string
      iconUri:
        type: string
      modelType:
        maximum: 3
        minimum: 1
        type: integer
      providerTemplateType:
        maximum: 7
        minimum: 1
        type: integer
    required:
    - config
    - displayName
    - modelType
    - providerTemplateType
    type: object
  sico-backend_internal_transport_http_dto_llmhubs.CreateModelRegistryResponse:
    properties:
      code:
        type: integer
      data:
        $ref: '#/definitions/sico-backend_internal_transport_http_dto_llmhubs.CreateModelRegistryData'
      msg:
        type: string
    type: object
  sico-backend_internal_transport_http_dto_llmhubs.DeleteModelRegistryRequest:
    properties:
      id:
        type: integer
    required:
    - id
    type: object
  sico-backend_internal_transport_http_dto_llmhubs.DeleteModelRegistryResponse:
    properties:
      code:
        type: integer
      msg:
        type: string
    type: object
  sico-backend_internal_transport_http_dto_llmhubs.ModelRegistryEntry:
    properties:
      agentId:
        type: string
      config:
        $ref: '#/definitions/structpb.Struct'
      createdAt:
        type: integer
      creatorUsername:
        type: string
      description:
        type: string
      displayName:
        type: string
      iconUri:
        type: string
      id:
        type: integer
      ioProfile:
        $ref: '#/definitions/structpb.Struct'
      isBuiltin:
        type: boolean
      modelKey:
        type: string
      modelType:
        type: integer
      providerTemplateType:
        type: integer
      status:
        type: integer
      updatedAt:
        type: integer
      updaterUsername:
        type: string
    type: object
  sico-backend_internal_transport_http_dto_llmhubs.RuntimeArtifact:
    properties:
      artifactType:
        type: string
      downloadUrl:
        type: string
      filename:
        type: string
      mimeType:
        type: string
      storageUri:
        type: string
    type: object
  sico-backend_internal_transport_http_dto_llmhubs.RuntimeGenerateRequest:
    properties:
      inputs:
        items:
          $ref: '#/definitions/sico-backend_internal_transport_http_dto_llmhubs.RuntimeInput'
        type: array
      instructions:
        type: string
      model:
        type: string
      options:
        $ref: '#/definitions/structpb.Struct'
      previousResponseId:
        type: string
      tools:
        items:
          $ref: '#/definitions/structpb.Struct'
        type: array
    required:
    - inputs
    type: object
  sico-backend_internal_transport_http_dto_llmhubs.RuntimeGenerateResponse:
    properties:
      artifacts:
        items:
          $ref: '#/definitions/sico-backend_internal_transport_http_dto_llmhubs.RuntimeArtifact'
        type: array
      code:
        type: integer
      msg:
        type: string
      outputs:
        items:
          $ref: '#/definitions/sico-backend_internal_transport_http_dto_llmhubs.RuntimeOutputItem'
        type: array
      payload:
        $ref: '#/definitions/structpb.Struct'
      trace:
        $ref: '#/definitions/sico-backend_internal_transport_http_dto_llmhubs.RuntimeTrace'
      usage:
        $ref: '#/definitions/sico-backend_internal_transport_http_dto_llmhubs.RuntimeUsage'
    type: object
  sico-backend_internal_transport_http_dto_llmhubs.RuntimeInput:
    properties:
      content:
        items:
          $ref: '#/definitions/sico-backend_internal_transport_http_dto_llmhubs.RuntimeInputContent'
        type: array
      role:
        type: string
    type: object
  sico-backend_internal_transport_http_dto_llmhubs.RuntimeInputContent:
    properties:
      actions:
        items:
          $ref: '#/definitions/structpb.Struct'
        type: array
      arguments:
        type: string
      callId:
        type: string
      detail:
        type: string
      fileBase64:
        type: string
      fileUrl:
        type: string
      imageBase64:
        type: string
      imageUrl:
        type: string
      mediaType:
        type: string
      name:
        type: string
      output:
        $ref: '#/definitions/structpb.Struct'
      result:
        type: string
      text:
        type: string
      type:
        type: string
    type: object
  sico-backend_internal_transport_http_dto_llmhubs.RuntimeOutputItem:
    properties:
      actions:
        items:
          $ref: '#/definitions/structpb.Struct'
        type: array
      arguments:
        type: string
      callId:
        type: string
      json:
        $ref: '#/definitions/structpb.Struct'
      name:
        type: string
      text:
        type: string
      type:
        type: string
    type: object
  sico-backend_internal_transport_http_dto_llmhubs.RuntimeStreamChunk:
    properties:
      code:
        type: integer
      delta:
        type: string
      finish_reason:
        type: string
      msg:
        type: string
      outputs:
        items:
          $ref: '#/definitions/sico-backend_internal_transport_http_dto_llmhubs.RuntimeOutputItem'
        type: array
      usage:
        $ref: '#/definitions/sico-backend_internal_transport_http_dto_llmhubs.RuntimeUsage'
    type: object
  sico-backend_internal_transport_http_dto_llmhubs.RuntimeTrace:
    properties:
      latencyMs:
        type: integer
      model:
        type: string
      providerTemplateType:
        type: integer
    type: object
  sico-backend_internal_transport_http_dto_llmhubs.RuntimeUsage:
    properties:
      completionTokens:
        type: integer
      promptTokens:
        type: integer
      totalTokens:
        type: integer
    type: object
  sico-backend_internal_transport_http_dto_project.AddProjectAssetData:
    properties:
      id:
        type: integer
      metaInfo:
        $ref: '#/definitions/sico-backend_internal_transport_http_dto_project.FileMetaInfo'
      sasUrl:
        type: string
      uri:
        type: string
    type: object
  sico-backend_internal_transport_http_dto_project.AddProjectAssetResponse:
    properties:
      code:
        type: integer
      data:
        $ref: '#/definitions/sico-backend_internal_transport_http_dto_project.AddProjectAssetData'
      msg:
        type: string
    type: object
...
  sico-backend_internal_transport_http_dto_project.DeleteProjectAssetResponse:
    properties:
      code:
        type: integer
      msg:
        type: string
    type: object
...
  sico-backend_internal_transport_http_dto_project.FileMetaInfo:
    properties:
      contentType:
        type: string
      fileExt:
        type: string
      fileName:
        type: string
      fileSize:
        type: integer
      fileType:
        type: string
    type: object
  sico-backend_internal_transport_http_dto_project.GetProjectAssetListData:
    properties:
      assets:
        items:
          $ref: '#/definitions/sico-backend_internal_transport_http_dto_project.ProjectAsset'
        type: array
      total:
        type: integer
    type: object
  sico-backend_internal_transport_http_dto_project.GetProjectAssetListResponse:
    properties:
      code:
        type: integer
      data:
        $ref: '#/definitions/sico-backend_internal_transport_http_dto_project.GetProjectAssetListData'
      msg:
        type: string
    type: object
...
  sico-backend_internal_transport_http_dto_project.GetProjectSASAssetData:
    properties:
      sas_url:
        type: string
      uri:
        type: string
    type: object
  sico-backend_internal_transport_http_dto_project.GetProjectSASAssetResponse:
    properties:
      code:
        type: integer
      data:
        $ref: '#/definitions/sico-backend_internal_transport_http_dto_project.GetProjectSASAssetData'
      msg:
        type: string
    type: object
...
  sico-backend_internal_transport_http_dto_project.ProjectAsset:
    properties:
      createdAt:
        type: integer
      creatorUsername:
        type: string
      extra:
        type: string
      id:
        type: integer
      objectKey:
        type: string
      projectId:
        type: string
      updatedAt:
        type: integer
    type: object
...
  structpb.Struct:
    properties:
      fields:
        additionalProperties:
          $ref: '#/definitions/structpb.Value'
        description: Unordered map of dynamically typed values.
        type: object
    type: object
  structpb.Value:
    properties:
      kind:
        description: "The kind of value.\n\nTypes that are valid to be assigned to
          Kind:\n\n\t*Value_NullValue\n\t*Value_NumberValue\n\t*Value_StringValue\n\t*Value_BoolValue\n\t*Value_StructValue\n\t*Value_ListValue"
    type: object
info:
  contact: {}
  description: Sico AI Agent Platform API
  title: Sico API
  version: "1.0"
paths:
...
  /api/sico/llm/models:
    delete:
      consumes:
      - application/json
      parameters:
      - description: Delete Model
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/sico-backend_internal_transport_http_dto_llmhubs.DeleteModelRegistryRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/sico-backend_internal_transport_http_dto_llmhubs.DeleteModelRegistryResponse'
      security:
      - BearerAuth: []
      tags:
      - llmhub
    post:
      consumes:
      - application/json
      parameters:
      - description: Create Model
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/sico-backend_internal_transport_http_dto_llmhubs.CreateModelRegistryRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/sico-backend_internal_transport_http_dto_llmhubs.CreateModelRegistryResponse'
      security:
      - BearerAuth: []
      tags:
      - llmhub
  /api/sico/llm/runtime/generate:
    post:
      consumes:
      - application/json
      parameters:
      - description: Runtime Generate
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/sico-backend_internal_transport_http_dto_llmhubs.RuntimeGenerateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/sico-backend_internal_transport_http_dto_llmhubs.RuntimeGenerateResponse'
      tags:
      - llmhub
  /api/sico/llm/runtime/generate/stream:
    post:
      consumes:
      - application/json
      parameters:
      - description: Runtime Generate
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/sico-backend_internal_transport_http_dto_llmhubs.RuntimeGenerateRequest'
      produces:
      - text/event-stream
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/sico-backend_internal_transport_http_dto_llmhubs.RuntimeStreamChunk'
      tags:
      - llmhub
  /api/sico/llm/sdk-examples:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      tags:
      - llmhub
  /api/sico/llm/source-slots:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      tags:
      - llmhub
...
  /api/sico/project/asset:
    delete:
      consumes:
      - application/json
      parameters:
      - in: query
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/sico-backend_internal_transport_http_dto_project.DeleteProjectAssetResponse'
      tags:
      - Project
    post:
      consumes:
      - multipart/form-data
      parameters:
      - description: Project ID
        in: formData
        name: project_id
        type: string
      - description: File to upload
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/sico-backend_internal_transport_http_dto_project.AddProjectAssetResponse'
      tags:
      - Project
  /api/sico/project/assets:
    get:
      consumes:
      - application/json
      parameters:
      - default: 1
        in: query
        minimum: 0
        name: page
        type: integer
      - default: 10
        in: query
        maximum: 100
        minimum: 1
        name: pageSize
        type: integer
      - in: query
        name: projectId
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/sico-backend_internal_transport_http_dto_project.GetProjectAssetListResponse'
      security:
      - BearerAuth: []
      tags:
      - Project
  /api/sico/project/sas_asset:
    get:
      consumes:
      - application/json
      parameters:
      - in: query
        name: uri
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/sico-backend_internal_transport_http_dto_project.GetProjectSASAssetResponse'
      tags:
      - Project
...
securityDefinitions:
  BearerAuth:
    description: 'Enter your bearer token in the format: Bearer <token>'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"